+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Release 2020-12-05-1714
+- [enh] switch to date tags for releases
+- [fix] join logsources into single string (with comma)
+- [fix] join referrer strings with semicolon (;) into single string
+- [add] message field containing Apache Common Log format representation of the session
+- [add] rubocop and solargraph to have a more pleasant vscode experience
+
 ## Release 0.1.1
 - [fix] only recalculate bytes for partial content when key extension is .mp3 (MP3 audio content)